BE/requirements.txt
Django==1.11
djangorestframework
django-cors-headers
psycopg2